remove dep cycle
